权限管理 超级管理员

+++ src/main/java/com/breech/extremity/dto/admin/UserRoleDTO.java
@@ -11,5 +11,5 @@ public class UserRoleDTO {
 
     private Long idUser;
 
-    private Long idRole;
+    private Integer idRole;
 }
